+++ acceptance-tests/src/test/java/tech/pegasys/ethsigner/tests/signing/PrivateTransactionAcceptanceTest.java
@@ -12,8 +12,6 @@
  */
 package tech.pegasys.ethsigner.tests.signing;
 
-import static io.netty.handler.codec.http.HttpResponseStatus.BAD_REQUEST;
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static tech.pegasys.ethsigner.core.jsonrpc.response.JsonRpcError.ENCLAVE_ERROR;
@@ -22,90 +20,16 @@ import static tech.pegasys.ethsigner.tests.dsl.Contracts.GAS_PRICE;
 import static tech.pegasys.ethsigner.tests.dsl.PrivateTransaction.RESTRICTED;
 
 import tech.pegasys.ethsigner.core.jsonrpc.response.JsonRpcErrorResponse;
-import tech.pegasys.ethsigner.tests.dsl.Account;
-import tech.pegasys.ethsigner.tests.dsl.DockerClientFactory;
+import tech.pegasys.ethsigner.tests.AcceptanceTestBase;
 import tech.pegasys.ethsigner.tests.dsl.PrivateTransaction;
-import tech.pegasys.ethsigner.tests.dsl.node.Node;
-import tech.pegasys.ethsigner.tests.dsl.node.NodeConfiguration;
-import tech.pegasys.ethsigner.tests.dsl.node.NodeConfigurationBuilder;
-import tech.pegasys.ethsigner.tests.dsl.node.PantheonNode;
-import tech.pegasys.ethsigner.tests.dsl.signer.Signer;
-import tech.pegasys.ethsigner.tests.dsl.signer.SignerConfiguration;
-import tech.pegasys.ethsigner.tests.dsl.signer.SignerConfigurationBuilder;
 import tech.pegasys.ethsigner.tests.dsl.signer.SignerResponse;
 import tech.pegasys.ethsigner.tests.signing.contract.generated.SimpleStorage;
 
-import java.io.IOException;
 import java.math.BigInteger;
-import java.net.URL;
 
-import com.github.dockerjava.api.DockerClient;
-import com.google.common.io.Resources;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
-public class PrivateTransactionAcceptanceTest {
-
-  private static final DockerClient DOCKER = new DockerClientFactory().create();
-
-  private static final String ENCLAVE_PUBLIC_KEY1 = "enclave_key1.pub";
-  private static final String ENCLAVE_PUBLIC_KEY2 = "enclave_key2.pub";
-
-  private Node ethNode;
-  private Signer ethSigner;
-
-  private Account richBenefactor() {
-    return ethSigner.accounts().richBenefactor();
-  }
-
-  private String enclavePublicKey1() {
-    return readResource(ENCLAVE_PUBLIC_KEY1);
-  }
-
-  private String enclavePublicKey2() {
-    return readResource(ENCLAVE_PUBLIC_KEY2);
-  }
-
-  @SuppressWarnings("UnstableApiUsage")
-  private String readResource(final String resourceName) {
-    final URL resource = Resources.getResource(resourceName);
-    try {
-      return new String(Resources.toString(resource, UTF_8).getBytes(UTF_8), UTF_8);
-    } catch (final IOException e) {
-      throw new RuntimeException("Unable to load resource " + resourceName);
-    }
-  }
-
-  @Before
-  public void setUp() {
-    Runtime.getRuntime().addShutdownHook(new Thread((this::tearDown)));
-
-    final NodeConfiguration nodeConfig =
-        new NodeConfigurationBuilder()
-            .withPrivacyEnabled()
-            .withPrivacyPublicKey(ENCLAVE_PUBLIC_KEY2)
-            .build();
-    ethNode = new PantheonNode(DOCKER, nodeConfig);
-    ethNode.start();
-    ethNode.awaitStartupCompletion();
-
-    final SignerConfiguration signerConfig = new SignerConfigurationBuilder().build();
-    ethSigner = new Signer(signerConfig, nodeConfig, ethNode.ports());
-    ethSigner.start();
-    ethSigner.awaitStartupCompletion();
-  }
-
-  @After
-  public void tearDown() {
-    if (ethNode != null) {
-      ethNode.shutdown();
-    }
-
-    if (ethSigner != null) {
-      ethSigner.shutdown();
-    }
-  }
+public class PrivateTransactionAcceptanceTest extends AcceptanceTestBase {
 
   @Test
   public void deployContract() {
@@ -118,14 +42,13 @@ public class PrivateTransactionAcceptanceTest {
             BigInteger.ZERO,
             SimpleStorage.BINARY,
             enclavePublicKey1(),
-            singletonList(enclavePublicKey2()),
+            singletonList(enclavePublicKey1()),
             RESTRICTED);
 
     final SignerResponse<JsonRpcErrorResponse> signerResponse =
-        ethSigner.privateContracts().submitExceptional(contract);
+        ethSigner().privateContracts().submitExceptional(contract);
     // We expect this to fail with enclave error as we don't have orion running. If rlp decode fails
     // then we would get a different error
-    assertThat(signerResponse.status()).isEqualTo(BAD_REQUEST);
     assertThat(signerResponse.jsonRpc().getError()).isEqualTo(ENCLAVE_ERROR);
   }
 }
