+++ acceptance-tests/src/test/resources/enclave_key.pub
@@ -1,0 +1,1 @@
+foobar
