--- acceptance-tests/src/test/resources/enclave_key1.pub
@@ -1,1 +1,0 @@
-A1aVtMxLCUHmBVHXoZzzBgPbW/wj5axDpW9X8l91SGo=
