--- acceptance-tests/src/test/resources/enclave_key2.pub
@@ -1,1 +1,0 @@
-Ko2bVqD+nNlNYL5EE7y3IdOnviftjiizpjRt+HTuFBs=
